app/requirements.txt
flask==2.2.2
pytest==7.1.2
pyyaml==6.0
prometheus_client==0.14.1
requests==2.28.1